opt(tikv/client-go): remove required status checks for golangci, race-test, and test
The Tide component will check for reported action run status when the
config for the repo has `skip-unknown-contexts` be false(default value).

With this change, we can make some simple PR which is only changes
markdown files or OWNERS files can be merged when it has commit message
includes '[SKIP CI]'.